% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_handle_cat_vars.R
\name{handle_cat_vars}
\alias{handle_cat_vars}
\title{handle_cat_vars}
\usage{
handle_cat_vars(kwargs)
}
\arguments{
\item{kwargs}{A list with key word arguments.}
}
\value{
Returns a list with two elements:
\itemize{
\item \code{params} The key word arguments without \code{cat_vars}.
\item \code{cat_vars} The vector \code{cat_vars}.
}
}
\description{
Helper function to handle categorical variables
}
\details{
This function is a utility function to separate the list element with the
names of the categorical variables from the key word arguments list to
be passed further on to \code{\link[kdry:dtr_matrix2df]{kdry::dtr_matrix2df()}}.
}
\examples{
handle_cat_vars(list(cat_vars = c("a", "b", "c"), arg1 = 1, arg2 = 2))

}
\seealso{
\code{\link[kdry:dtr_matrix2df]{kdry::dtr_matrix2df()}}
}
