% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_lm.R
\name{LearnerLm}
\alias{LearnerLm}
\title{LearnerLm R6 class}
\description{
This learner is a wrapper around [stats::lm()] in order to perform a
  linear regression. There is no implementation for tuning
  parameters, thus the only experiment to use `LearnerLm` for is
  [mlexperiments::MLCrossValidation]
}
\details{
Implemented methods:
\describe{
\item{`$fit`}{To fit the model.}
\item{`$predict`}{To predict new data with the model.}
}
}
\examples{
LearnerLm$new()


## ------------------------------------------------
## Method `LearnerLm$new`
## ------------------------------------------------

LearnerLm$new()

}
\seealso{
[stats::lm()]

[stats::lm()]
}
\section{Super class}{
\code{\link[mlexperiments:MLLearnerBase]{mlexperiments::MLLearnerBase}} -> \code{LearnerLm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerLm-new}{\code{LearnerLm$new()}}
\item \href{#method-LearnerLm-clone}{\code{LearnerLm$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="bayesian_scoring_function"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-bayesian_scoring_function'><code>mlexperiments::MLLearnerBase$bayesian_scoring_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="cross_validation"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-cross_validation'><code>mlexperiments::MLLearnerBase$cross_validation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="fit"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-fit'><code>mlexperiments::MLLearnerBase$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlexperiments" data-topic="MLLearnerBase" data-id="predict"><a href='../../mlexperiments/html/MLLearnerBase.html#method-MLLearnerBase-predict'><code>mlexperiments::MLLearnerBase$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerLm-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerLm-new}{}}}
\subsection{Method \code{new()}}{
Create a new `LearnerLm` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerLm$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This learner is a wrapper around [stats::lm()] in order to perform a
  linear regression. There is no implementation for tuning
  parameters, thus the only experiment to use `LearnerLm` for is
  [mlexperiments::MLCrossValidation]
}

\subsection{Returns}{
A new `LearnerLm` R6 object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{LearnerLm$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerLm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerLm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerLm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
